show build timestamp instead of OSTree timestamp
Build timestamps represent when each build was created, whereas OSTree
timestamps aren't necessarily unique across builds. Let's reflect that
by dropping the OSTree timestamp and instead having a relative build
timestamp (with full date as tooltip).

In practice, the majority of builds will result in a new OSTree, so the
build timestamp is also roughly the OSTree commit timestamp. In the
future, we can add it back if we find it useful.