Support new multi-arch bucket layout
As described in https://github.com/coreos/coreos-assembler/pull/580.

The word "support" is used loosely here. Properly supporting and
presenting multi-arch info would require reworking the UI with e.g. a
drop-down or selectors for choosing which arch's artifacts to download.

For now, the layout doesn't change: in the non-legacy case, we simply
show the artifacts related to the first arch (which right now is the
only arch: x86_64).