Drop Content Checksum, add Parent Commit
The content checksum was initially intended to track content across
promotions with a grafting strategy. But we're not going to use that
approach anymore, at least for the time being. Instead, we'll always be
specifying the latest release on a stream as the parent of all the
builds of that stream. Let's highlight it so we can sanity-check it.