browser: don't print "Images:" if none exist
E.g. if the pipeline is configured to build only the oscontainer, there
won't be any images to print. It looks funny to still print the heading.
Since at least the OSTree commit is always present, we just need to
check if the size of `.images` is greater than 1.

We need to do this outside of the templating because there's no access
to the `Object` type there.