browser: rework pushed containers display
Abbreviate the digest pullspec so it fits better on screens, but keep it
in full in the link copy.

Print a pullspec for each tag pushed as an way to see which tags it was
pushed as and to make copying the associated pullspec easy.

Move the extensions container lower down since it's not as important as
the base oscontainer and e.g. package listing.