browser: tolerate missing coreos-assembler metadata
The current builds done with the buildah path are missing some of the
usual git metadata for the source repo and cosa repo.

We do want that information in there (though for the source config repo,
using the standard OCI org.opencontainers labels, which will end up in
`meta.json` already since we now capture all labels there), but for now
at least let's tolerate that information missing so builds show up.